hw1/requirements.txt
Package    Version
---------- -------
cffi       1.11.5 
pip        19.0.1 
pycparser  2.19   
PyNaCl     1.3.0  
setuptools 40.8.0 
six        1.12.0 
wheel      0.32.3 
